UI Tweaks and basic input functionality
- Added Create account button
- Adjusted UI
- Added basic script functionality to read the input from the player

The typed input is cast to the login script but not currently used in any capacity.

Next steps would be taking the player to the relevant canvas screen after the player logs in/creates an account, or requesting the player re-enter the info if the login request is invalid or the account name is taken.